requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
azure-common==1.1.28
azure-core==1.33.0
azure-search-documents==11.5.2
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
gotrue==2.12.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jwt==1.3.1
lxml==5.3.2
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.2.4
openai==1.70.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
pluggy==1.6.0
postgrest==1.0.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
PyMuPDF==1.25.5
pyngrok==7.2.11
PyPDF2==3.0.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
XlsxWriter==3.2.2
yarl==1.20.0
gunicorn
